+++ front/src/components/Login/Login.tsx
@@ -1,29 +1,17 @@
 import React, { FC, useState } from "react";
 import { Link, useNavigate } from "react-router-dom";
-import { useForm } from "react-hook-form";
-import { zodResolver } from "@hookform/resolvers/zod";
 import Cookies from "js-cookie";
-import { FaEye } from "react-icons/fa";
-import { FaEyeSlash } from "react-icons/fa";
 
-import { loginValidationSchema } from "../../utils/validationSchema";
-import NeutralButton from "../Buttons/NeutralButton";
 import { LoginParams } from "../../types/index";
 import { login } from "../../lib/api/auth";
 import { useAuthContext } from "../../context/AuthContext";
 import PageHelmet from "../PageHelmet";
 import { toast } from "react-toastify";
 import LoadingSpinner from "../Loadings/LoadingSpinner";
+import LoginForm from "./LoginForm";
 
 const Login: FC = () => {
-  const {
-    register,
-    handleSubmit,
-    formState: { errors },
-  } = useForm<LoginParams>({ resolver: zodResolver(loginValidationSchema) });
-
   const [loading, setLoading] = useState<boolean>(false);
-  const [isRevealPassword, setIsRevealPassword] = useState<boolean>(false);
   const navigate = useNavigate();
   const { setIsSignedIn, setCurrentUser } = useAuthContext();
 
@@ -47,12 +35,6 @@ const Login: FC = () => {
     setLoading(false);
   };
 
-  const buttonText = "ログイン";
-
-  const handleRevealPassword = () => {
-    setIsRevealPassword(!isRevealPassword);
-  };
-
   return (
     <>
       <PageHelmet title="ログイン" />
@@ -60,41 +42,7 @@ const Login: FC = () => {
       <div className="flex flex-col items-center h-screen mt-20">
         <div className="w-96 bg-white rounded p-6 shadow-xl">
           <h2 className="text-2xl text-center font-bold mb-5 text-gray-800">ログイン画面</h2>
-          <form onSubmit={handleSubmit(onSubmit)}>
-            <div className="mb-2">
-              <label className="mb-2 text-sm" htmlFor="email">
-                メールアドレス
-              </label>
-              <input
-                className="w-full px-3 py-2 border rounded-md outline-none"
-                id="email"
-                type="email"
-                {...register("email")}
-              />
-              {errors.email && <span className="text-error">{errors.email.message}</span>}
-            </div>
-            <div className="mb-5 relative">
-              <label className="mb-2 text-sm" htmlFor="password">
-                パスワード
-              </label>
-              <input
-                className="w-full px-3 py-2 border rounded-md outline-none"
-                id="password"
-                type={isRevealPassword ? "text" : "password"}
-                {...register("password")}
-              />
-              <span
-                onClick={handleRevealPassword}
-                className="absolute right-3 top-2/3 transform -translate-y-1/2"
-              >
-                {isRevealPassword ? <FaEyeSlash /> : <FaEye />}
-              </span>
-              {errors.password && <span className="text-error">{errors.password.message}</span>}
-            </div>
-            <div className="flex justify-center">
-              <NeutralButton buttonText={buttonText} />
-            </div>
-          </form>
+          <LoginForm onSubmit={onSubmit} />
         </div>
         <Link to="/password-reset-request" className="mt-6 link link-hover">
           パスワードをお忘れの方はこちら
